ci: make E2E folder based
